+++ src/eval/bc/stencils/abi.h
@@ -334,6 +334,19 @@ extern mino_val_t **mino_jit_make_lazy_slow(mino_state_t *S,
                                              unsigned a,
                                              mino_bc_fn_t *bc,
                                              unsigned bx);
+extern mino_val_t **mino_jit_protocol_call_cached_slow(mino_state_t *S,
+                                                        mino_val_t **regs,
+                                                        unsigned a,
+                                                        unsigned argn,
+                                                        unsigned ret,
+                                                        mino_bc_fn_t *bc,
+                                                        unsigned slot_idx);
+extern mino_val_t  *mino_jit_protocol_tailcall_cached_slow(mino_state_t *S,
+                                                            mino_val_t **regs,
+                                                            unsigned a,
+                                                            unsigned argn,
+                                                            mino_bc_fn_t *bc,
+                                                            unsigned slot_idx);
 extern mino_val_t **mino_jit_push_env_slow(mino_state_t *S,
                                             mino_val_t **regs);
 extern mino_val_t **mino_jit_pop_env_slow(mino_state_t *S,